Adding new class structure for maps api and driver
