+++ PythonGoogleMapsAPI/GoogleMapsAPI/GoogleMapsApi/MainGoogleMapsApi.py
@@ -1,0 +1,34 @@
+import ElevationMap
+import StaticMapsApi
+import MapTilesApi
+
+class MainGoogleMapsApi:
+    # Add Constructor
+    def __init__(self):
+        # Initialize the instances of the other classes
+        self.= Calculator()
+        self.greeter = Greeter()
+        self.logger = Logger()
+
+    # Define a run class to drive program execution
+    def run(self):
+        # General example usage for StaticMapsApi
+        api_key = "AIzaSyAHFnNvs2uAGYs7ur-uETNR-6Kle2VFIIo"
+        client = googlemaps.Client(api_key)
+
+        download_static_map(
+            client,
+            size=(400, 400),
+            center=(63.259591, -144.667969),  # Coordinates of the center of the map
+            zoom=6,
+            scale=2,
+            format="png",
+            maptype="roadmap",
+            markers=[(62.107733, -145.541936), "Delta+Junction,AK"]
+        )
+        
+
+# Instantiate the driver and run it
+if __name__ == "__main__":
+    driver = Driver()
+    driver.run()
